logic working, layouts broken

+++ core/network/eskills/src/main/java/com/appcoins/wallet/core/network/eskills/model/User.kt
@@ -13,5 +13,5 @@ class User(
   val status: UserStatus,
 
   @SerializedName("score")
-  var score: Int
+  var score: Long
 )
